rename roundoff example

+++ A15-roundoff.c
@@ -8,8 +8,8 @@
         http://docs.oracle.com/cd/E19957-01/806-3568/ncg_goldberg.html
 */
 int main(){
-    float a;
-    float b;
+    double a;
+    double b;
 
     for (int i=1; i<=50; i++){
         a += 0.9;
